Rename some functions, variables, rearrange some folders, add logout functionality
