+++ auth.ts
@@ -25,7 +25,8 @@ const tokenProvider = new TokenProvider(
 
 export const cleanUpSession = () => {
   tokenProvider.invalidateTokenInfo();
-  return clearCookie('auth');
+  clearCookie('auth');
+  clearCookie('isAuth');
 };
 
 export const clientLogin = async (
@@ -41,7 +42,7 @@ export const clientLogin = async (
   await tokenProvider.invalidateTokenInfo();
   return apolloClient
     .resetStore()
-    .then(() => setCookie('isAuthenticated', true))
+    .then(() => setCookie('isAuth', true))
     .catch(error => {
       // eslint-disable-next-line no-console
       console.error('Error on cache reset during login', error);
@@ -49,9 +50,13 @@ export const clientLogin = async (
     });
 };
 
-export const clientLogout = (apolloClient: ApolloClientType, redirect) => {
+export const clientLogout = (
+  apolloClient: ApolloClientType,
+  redirect: () => void
+) => {
   cleanUpSession();
-  return redirect().then(() => apolloClient.resetStore());
+  apolloClient.resetStore();
+  return redirect();
 };
 
 export const getTokenInfo = async () => {
