+++ public/locales/de/footer.json
@@ -1,5 +1,5 @@
 {
-  "titleOne":  "Hilfe",
+  "titleOne": "Hilfe",
   "faq": "FAQ",
   "deliveryAndReturn": "Lieferungen & Retouren",
   "requestReturn": "Fordern Sie eine Rücksendung an",
@@ -22,7 +22,7 @@
   "termsOfUse": "Nutzungsbedingungen",
   "privacyDeclaration": "Datenschutzerklärung",
   "cookiePolicy": "Cookie-Richtlinie",
-  "year": "Jahr",
+  "year": "2021",
   "crown": "Crown",
   "rightsReserved": "Alle Rechte vorbehalten"
 }
